+++ runtime/core/xnix_runtime_core_cli.c
@@ -2830,6 +2830,8 @@ print_probe(void)
   fputs("\"notification_plan_count\":4,", stdout);
   fputs("\"tray_status_plan_owner\":\"c\",", stdout);
   fputs("\"tray_status_plan_count\":1,", stdout);
+  fputs("\"krunner_query_plan_owner\":\"c\",", stdout);
+  fputs("\"krunner_query_plan_count\":1,", stdout);
   fputs("\"portal_request_plan_owner\":\"c\",", stdout);
   fputs("\"portal_request_plan_count\":7,", stdout);
   fputs("\"write_methods\":", stdout);
@@ -2892,5 +2894,6 @@ print_write_gate(const char *method_name)
 #include "xnix_runtime_core_cli_file_association.inc"
 #include "xnix_runtime_core_cli_notification.inc"
 #include "xnix_runtime_core_cli_tray_status.inc"
+#include "xnix_runtime_core_cli_krunner_query.inc"
 #include "xnix_runtime_core_cli_portal_request.inc"
 #include "xnix_runtime_core_cli_dispatch.inc"